Added aside section
Using grid line placement, we now only extend featured article until the 2nd to last column generated.
The remaining column is now taken up by the trending/popular/latest sidebar.
The aside content uses a nested grid to spread each section (or row) to the required height to be equal to the height of the featured article row, or row 2 of the main grid.
Maximum width of the main grid has been expanded to 1200px up from 1000px